+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -99,6 +100,11 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if ($childCategory > 0) {
+            return response(['status' => 'error', 'message' => 'This items contain sub items for delete!']);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted successfuly!']);
